backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
scikit-learn
pandas
numpy
psycopg2-binary
python-dotenv
gunicorn
whitenoise
dj-database-url
